Extra checkers (#71)
* Add whitespace checker to golangci

There are plenty of irregularites where some functions have unnecessary
whitespaces while others don't. This is an attempt to make it more
consistent.

* Add godot to golangci

This makes sure that all comments end in a period, which makes it
easier to remember to write full sentences.

* Add contextcheck to golangci

Also base the example server on the incoming request's context.