Make local evaluation non-blocking (#73)
* Simplify creation of error message

* Enable missing local evaluation setting in test

This will panic in the next commit if left untouched, and it appears
like it was a misstake to not have local evaluation enabled to begin
with.

* Refactor how local evaluation is used

If local evaluation is enabled then the primary functions used to
fetch the flags will not reach out to the API (which could block for
several seconds), but instead use the environmet or directly go to
the default flag handler.

This change will make so that execution of the program will always
be snappy, even in the case of flaky network connection. Also the
*FromAPI() functions will actually return an error in case they fail
to get the latest information from the API instead of using the
default flag handler.

While these aren't _really_ breaking changes, they significantly
alter how and when errors are returned and I would suggest this is
a major release change.