Allow defaultFlagHandler to return an error
Currently the signature of the default Flag handler only allows
for a flag to be returned, with no way to signal back in case
something goes wrong when trying to find the requested feature.

This is different behavior compared to how it works when querying
the FlagsmithAPI, which is why we now try to align it.

With this change we would be able to write a handler like this:

  func improvedDefaultHandler(feature string) (Flag, error) {
    switch feature {
    case "exists":
      return Flag{FeatureName: "exists", Value: "nice" }, nil
    default:
      return Flag{}, fmt.Errorf("feature %q does not exist", feature)
    }
  }