Only clear analytics cache in case of successful transmission
When analytics is enabled a background process will transmit the
locally stored cache on a set interval, however, in the previous
solution we would empty this cache even though the sending failed.
This would cause us to lose data without any way of preventing it.

In this change we only clear the cache in case we know it has reached
the server successfully.

Solves #74